Time: 1 ms (2.62%), Space: 42.1 MB (53.03%) - LeetHub

+++ 0319-bulb-switcher/0319-bulb-switcher.java
@@ -1,7 +1,7 @@
 class Solution {
    public int bulbSwitch(int n) {
         int count=0;
-        for(int i=1;i*i<=n;i++){
+        for(double i=1;i*i<=n;i++){
             if(i*i<=n) count++;
         }
         return count;
